% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mail.R
\name{attachments}
\alias{attachments}
\title{attachments}
\usage{
attachments(mail, path, name)
}
\arguments{
\item{mail}{mail object from package}

\item{path}{file path to attach}

\item{name}{file name. default is path's file name}
}
\description{
attachments
}
