% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_check}
\alias{auth_check}
\title{Check for a working API secret to authorize with Sendgrid.}
\usage{
auth_check()
}
\value{
TRUE/FALSE check work fine return TRUE.
}
\description{
Check for a working API secret to authorize with Sendgrid.
}
