% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_key}
\alias{auth_key}
\title{Returns the \code{SENDGRID_KEY} environment variable for unlocking non-system default keyring on Linux systems.}
\usage{
auth_key()
}
\value{
\code{chr} The \code{SENDGRID_KEY} environment variable value
}
\description{
Returns the \code{SENDGRID_KEY} environment variable for unlocking non-system default keyring on Linux systems.
}
