% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_secret}
\alias{auth_secret}
\title{Retrieve the SendGrid API Secret for authorization with the Sendgrid API}
\usage{
auth_secret()
}
\value{
The sendgrid API Secret
}
\description{
Retrieve the SendGrid API Secret for authorization with the Sendgrid API
}
