% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_set}
\alias{auth_set}
\title{Set API secret for authorization with the Sendgrid API.}
\usage{
auth_set(apisecret = Sys.getenv("SENDGRID_SECRET"))
}
\arguments{
\item{apisecret}{sendgrid api secret. Can be set as Environment variable \code{SENDGRID_SECRET} or provided interactively via prompt input.}
}
\value{
None
}
\description{
The \code{apisecret} must have read access or higher for the \verb{API Keys} permission. See the Settings > API Keys menu in sendgrid to create & modify key permissions.
}
