% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mail.R
\name{bcc}
\alias{bcc}
\title{bcc}
\usage{
bcc(mail, email, name = "")
}
\arguments{
\item{mail}{mail object from package}

\item{email}{email address}

\item{name}{name for email address}
}
\description{
bcc
}
