% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mail.R
\name{body}
\alias{body}
\title{body}
\usage{
body(sg_mail, body, type = "text/html")
}
\arguments{
\item{sg_mail}{(required)mail object from package}

\item{body}{(required)mail content html support.}

\item{type}{content type. text/html is default.}
}
\value{
sg_mail class with body content.
}
\description{
body
}
\examples{
mail() \%>\%
  body("mrchypark@gmail.com")
}
