% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mail.R
\name{content}
\alias{content}
\title{content}
\usage{
content(sg_mail, content, type = "text/html")
}
\arguments{
\item{sg_mail}{(required)mail object from package}

\item{content}{(required)mail content html support.}

\item{type}{content type. text/html is default.}
}
\description{
content
}
