% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mail.R
\name{dynamic_template_data}
\alias{dynamic_template_data}
\title{Add dynamic tempalte data}
\usage{
dynamic_template_data(sg_mail, tbl)
}
\arguments{
\item{sg_mail}{(required) mail object from package}

\item{tbl}{A dataframe. Column names will become the key names for your substitutions. The values will become the values they're assinged. (See \url{https://docs.sendgrid.com/ui/sending-email/how-to-send-an-email-with-dynamic-transactional-templates})}
}
\value{

}
\description{
Add dynamic tempalte data
}
\examples{
sub_tbl <-
  tibble(
    first_name = "Amanda",
    link = "foo"
  )

mail() \%>\%
  dynamic_template_data(sub_tbl)
}
