% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendgrid.R
\name{email_person}
\alias{email_person}
\title{Constructs list with email and name of person}
\usage{
email_person(email, name)
}
\arguments{
\item{email}{email address}

\item{name}{name}
}
\value{
list of length one
}
\description{
Used to construct the email payload. See \link{make_email_payload} and \link{sendgrid_send}
}
\seealso{
Other email functions: \code{\link{make_email_payload}},
  \code{\link{sendgrid_send}}
}
