% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mail.R
\name{from}
\alias{from}
\title{from}
\usage{
from(sg_mail, email, name = "")
}
\arguments{
\item{sg_mail}{(required)mail object from package}

\item{email}{(required)email address}

\item{name}{name for email address}
}
\value{
sg_mail class with from mail address.
}
\description{
from
}
\examples{
mail() \%>\%
  from("mrchypark@gmail.com")
}
