% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\docType{data}
\name{kr}
\alias{kr}
\title{Environment to store keyring related variables}
\format{
An object of class \code{environment} of length 3.
}
\usage{
kr
}
\description{
Environment to store keyring related variables
}
\keyword{datasets}
