% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendgrid.R
\name{make_email_payload}
\alias{make_email_payload}
\title{Constructs email payload from emails, subject and body}
\usage{
make_email_payload(to, subject, body, from, reply_to = from)
}
\arguments{
\item{to}{\link{email_group}}

\item{subject}{character vector with subject of message}

\item{body}{character vector with body of message}

\item{from}{\link{email_person}}

\item{reply_to}{\link{email_person}}
}
\value{
list with payload in the format required by \link{sendgrid_send}
}
\description{
Once the payload is constructed, you can use \link{sendgrid_send} to send the message
}
\seealso{
Other email functions: \code{\link{email_person}},
  \code{\link{sendgrid_send}}
}
