% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.sg_mail}
\alias{print.sg_mail}
\title{print sg_mail class}
\usage{
\method{print}{sg_mail}(x, ...)
}
\arguments{
\item{x}{mail object from package}

\item{...}{print params}
}
\description{
print sg_mail class
}
