% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendgrid-package.R
\docType{package}
\name{sendgrid-package}
\alias{sendgrid-package}
\alias{sendgrid}
\title{Connect to sendgrid from R}
\description{
Connect to sendgrid from R
}
\author{
Andrie de Vries \email{apdevries@gmail.com}
}
\keyword{package}
