% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendgrid.R
\name{sendgrid_send}
\alias{sendgrid_send}
\title{Sends mail using sendgrid API v3}
\usage{
sendgrid_send(payload, api_key)
}
\arguments{
\item{payload}{email payload, see \link{make_email_payload}}

\item{api_key}{character string, obtained as API key in sendgrid app}
}
\value{
Called for its side effect of sending the payload as an email using the sendgrid API. Returns NULL if successful
}
\description{
Sends mail using sendgrid API v3
}
\references{
https://sendgrid.api-docs.io/v3.0/mail-send
}
\seealso{
Other email functions: \code{\link{email_person}},
  \code{\link{make_email_payload}}
}
