% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendgridr-package.R
\docType{package}
\name{sendgridr-package}
\alias{sendgridr}
\alias{sendgridr-package}
\title{sendgridr: Mail Sender Using Sendgrid Service}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Send email using sendgrid mail api(v3).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/sendgridr}
  \item Report bugs at \url{https://github.com/mrchypark/sendgridr/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (blog, <mrchypark.githubi.io>)

}
\keyword{internal}
