% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mail.R
\name{address}
\alias{address}
\alias{to}
\alias{cc}
\alias{bcc}
\title{address}
\usage{
address(locate)

to(sg_mail, email, name = "")

cc(sg_mail, email, name = "")

bcc(sg_mail, email, name = "")
}
\arguments{
\item{locate}{where to set mail address}

\item{sg_mail}{(required)mail object from package}

\item{email}{(required)email address}

\item{name}{(optional)name for email address.}
}
\description{
mail address

to(), cc(), bcc() is for set email address to sg_mail class
}
