% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mail.R
\name{address}
\alias{address}
\alias{to}
\alias{cc}
\alias{bcc}
\title{set address to sg_mail class}
\usage{
to(sg_mail, email, name = "")

cc(sg_mail, email, name = "")

bcc(sg_mail, email, name = "")
}
\arguments{
\item{sg_mail}{(required)mail object from package}

\item{email}{(required)email address}

\item{name}{(optional)name for email address.}
}
\value{
sg_mail class with mail address.
}
\description{
to(), cc(), bcc() is for set email address to sg_mail class
}
\examples{
mail() \%>\%
  to("mrchypark@gmail.com")

mail() \%>\%
  cc("mrchypark@gmail.com")

mail() \%>\%
  bcc("mrchypark@gmail.com")
}
