% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mail.R
\name{attachments}
\alias{attachments}
\title{attachments}
\usage{
attachments(sg_mail, path, name, content_id)
}
\arguments{
\item{sg_mail}{(required)mail object from package}

\item{path}{(required)file path to attach}

\item{name}{file name. default is path's file name}

\item{content_id}{content id(cid). default is `content_id``}
}
\description{
attachments
}
