% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mail.R
\name{attachments}
\alias{attachments}
\title{attachments}
\usage{
attachments(sg_mail, path, name, content_id)
}
\arguments{
\item{sg_mail}{(required)mail object from package}

\item{path}{(required)file path to attach}

\item{name}{file name. default is path's file name}

\item{content_id}{content id. default is Null.}
}
\value{
sg_mail class with attachments.
}
\description{
attachments
}
\examples{
path <- system.file("extdata", "test.Rmd", package = "sendgridr")
mail() \%>\%
  attachments(path)

}
