% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mail.R
\name{attachments}
\alias{attachments}
\title{attachments}
\usage{
attachments(sg_mail, path, name)
}
\arguments{
\item{sg_mail}{(required)mail object from package}

\item{path}{(required)file path to attach}

\item{name}{file name. default is path's file name}
}
\value{
sg_mail class with attachments.
}
\description{
attachments
}
\examples{
\dontrun{
mail() \%>\%
  attachments("sendgridr.docx")
}
}
