% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_check}
\alias{auth_check}
\title{Check API key for auth.}
\usage{
auth_check()
}
\value{
TRUE/FALSE check work fine return TRUE.
}
\description{
Check SENDGRID_API env value is set.
If any value from SENDGRID_API detected, return TRUE.
Function works using auth_check_zero(), auth_check_dummy(), auth_check_work().
}
