% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_set}
\alias{auth_set}
\title{Set API key for auth.}
\usage{
auth_set(apikey)
}
\arguments{
\item{apikey}{sendgrid api key. If key has value, just use as api key. If without param, function call interactive prompt input.}
}
\value{
None
}
\description{
Set API key for auth.
}
