% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_set}
\alias{auth_set}
\title{Set API key for auth.}
\usage{
auth_set(force = FALSE)
}
\arguments{
\item{force}{force set api key process. defualt is FALSE.}
}
\value{
None
}
\description{
If run code, open .Renviron file for set api key.
}
