% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mail.R
\name{dynamic_template_data}
\alias{dynamic_template_data}
\title{Add dynamic template data}
\usage{
dynamic_template_data(sg_mail, lst)
}
\arguments{
\item{sg_mail}{(required) mail object from package}

\item{lst}{A key-value list for template data. (See \url{https://docs.sendgrid.com/ui/sending-email/how-to-send-an-email-with-dynamic-transactional-templates})}
}
\value{

}
\description{
\code{template_id} must be included for this data to be applied.
}
\examples{
data_lst <-
  list(
    first_name = "Amanda",
    link = "foo"
  )

mail() \%>\%
  dynamic_template_data(data_lst)
}
