% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.sg_mail}
\alias{print.sg_mail}
\title{Print for sg_mail class}
\usage{
\method{print}{sg_mail}(x, ...)
}
\arguments{
\item{x}{sg_mail class object}

\item{...}{pass for default print. But not use in sg_mail class.}
}
\description{
Print for sg_mail class
}
