% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendgridr-package.R
\docType{package}
\name{sendgridr-package}
\alias{sendgridr}
\alias{sendgridr-package}
\title{sendgridr: Mail Sender Using 'Sendgrid' Service}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Send email using 'Sendgrid' \url{https://sendgrid.com/} mail API(v3) \url{https://docs.sendgrid.com/api-reference/how-to-use-the-sendgrid-v3-api/authentication}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/sendgridr}
  \item Report bugs at \url{https://github.com/mrchypark/sendgridr/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (\href{https://orcid.org/0000-0001-6474-2570}{ORCID})

}
\keyword{internal}
