% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mail.R
\name{subject}
\alias{subject}
\title{subject}
\usage{
subject(sg_mail, subject)
}
\arguments{
\item{sg_mail}{(required)mail object from package}

\item{subject}{(required)mail subject}
}
\value{
sg_mail class with subject.
}
\description{
subject
}
\examples{
mail() \%>\%
  subject("mrchypark@gmail.com")
}
